feat: 공통 컴포넌트 생성 및 파일 분리
