+++ src/components/base/Card.jsx
@@ -5,8 +5,9 @@ import size from '../../utils/size';
 import { body1, body3 } from '../../styles/fonts';
 import Check from '../../assets/icons/Check';
 import { colors } from '../../styles/colors';
+import FolderBox from './FolderBox';
 
-const Card = ({ title = '', numberOfLinks = 0, onPress, isSelected, isOnlyTitle }) => {
+const Card = ({ title = '', numberOfLinks = 0, onPress, isSelected }) => {
   const handleSliceText = () => {
     return title.length > 16 ? `${title.slice(0, 16)}...` : title;
   };
@@ -17,7 +18,11 @@ const Card = ({ title = '', numberOfLinks = 0, onPress, isSelected, isOnlyTitle 
 
   return (
     <CardWrapper onPress={onPress}>
-      <ImageBox source={require('../../assets/images/logo.png')} />
+      <View style={{ marginRight: size.width * 16 }}>
+        <FolderBox iconSize={64}>
+          <ImageBox source={require('../../assets/images/folder-yellow.png')} />
+        </FolderBox>
+      </View>
       <TextWrapper>
         <Title>{handleSliceText()}</Title>
         <Subtitle>{handleSliceLink()}개의 링크</Subtitle>
@@ -35,7 +40,7 @@ export default Card;
 
 const CardWrapper = styled(TouchableOpacity)`
   width: '100%';
-  height: ${size.height * 96};
+  height: ${size.height * 96}px;
   flex-direction: row;
   align-items: center;
   background-color: ${colors.bg[400]};
@@ -43,9 +48,8 @@ const CardWrapper = styled(TouchableOpacity)`
 `;
 
 const ImageBox = styled(Image)`
-  width: ${size.width * 64}px;
-  height: ${size.height * 64}px;
-  margin-right: ${size.width * 16}px;
+  width: ${size.width * 29.222}px;
+  height: ${size.height * 26.814}px;
 `;
 
 const TextWrapper = styled(View)`
@@ -54,13 +58,13 @@ const TextWrapper = styled(View)`
 
 const Title = styled(Text)`
   font-family: ${body1.medium.fontFamily};
-  font-size: ${body1.medium.fontSize};
+  font-size: ${body1.medium.fontSize}px;
   color: white;
 `;
 
 const Subtitle = styled(Text)`
   font-family: ${body3.medium.fontFamily};
-  font-size: ${body3.medium.fontSize};
+  font-size: ${body3.medium.fontSize}px;
   color: white;
 `;
