+++ src/components/base/Toast.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const Toast = () => {
+  return <div></div>;
+};
+
+export default Toast;
